expected_variance

Calculate expected variance in response variable(s)

Usage

expected_variance(squid)

Arguments

squid

A squid object

Value

A data.frame with the data structure

Details

Calculated the expect variance from the simulation parameters. Has several limitaitons. 1. Doesn't work when 'model' is specific in simulate_population 2. Assumes random factors are balanced - unbalanced designs will change the observed variance 3. Will be inaccurate with transformed variables (i.e. if functions are specified) 4. Doesn't deal well with three way interactions and over, unless they are the same variable (i.e. polynomials) 5. Doesn't account for covariance between interaction terms, e.g. if there is are two interaction terms rain:temp and wind:temp, they will covary, but this additional variance in the response is not calculated

Author

Joel Pick - joel.l.pick@gmail.com

Examples

# simple data structure with 5 'individuals' and 2 observations per individual
make_structure(structure="individual(5)", repeat_obs=2)
#>    individual
#> 1           1
#> 2           1
#> 3           2
#> 4           2
#> 5           3
#> 6           3
#> 7           4
#> 8           4
#> 9           5
#> 10          5

# nested data structure with 2 sexes, 5 individuals per sex
# and 2 observations per individual
make_structure(structure="sex(2)/individual(5)", repeat_obs=2)
#>    sex individual
#> 1    1          1
#> 2    1          1
#> 3    1          2
#> 4    1          2
#> 5    1          3
#> 6    1          3
#> 7    1          4
#> 8    1          4
#> 9    1          5
#> 10   1          5
#> 11   2          6
#> 12   2          6
#> 13   2          7
#> 14   2          7
#> 15   2          8
#> 16   2          8
#> 17   2          9
#> 18   2          9
#> 19   2         10
#> 20   2         10

# crossed data structure with 5 individuals in 2 treatments 
# and 2 observations per individual and treatment combination
make_structure(structure="treatment(2) + individual(5)", repeat_obs=1)
#>    treatment individual
#> 1          1          1
#> 2          1          2
#> 3          1          3
#> 4          1          4
#> 5          1          5
#> 6          2          1
#> 7          2          2
#> 8          2          3
#> 9          2          4
#> 10         2          5